requirements.txt
fastapi==0.135.1
uvicorn==0.36.0
pydantic==2.10.5
requests==2.32.3
python-dotenv==1.0.1
